pyproject.toml
[project]
name = "citedy-reddit-writer"
version = "0.1.1"
description = "Reddit listings + Citedy Agent API autopilot (no scout credits). Optional interactive setup."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27,<1",
  "pyyaml>=6,<7",
]

[project.scripts]
citedy-reddit-run = "citedy_reddit_writer.run_once:main"
citedy-reddit-setup = "citedy_reddit_writer.setup_wizard:main"

[project.urls]
Homepage = "https://github.com/citedy/citedy-reddit-writer"
Repository = "https://github.com/citedy/citedy-reddit-writer"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["citedy_reddit_writer"]
